+++ src/plugins/csp_configuration_provider/server/routes/index.ts
@@ -1,8 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { IRouter } from '../../../../core/server';
-
-export function defineRoutes(router: IRouter) {}